Rework shutdown logic again
It's almost like concurrent programming is hard...

The use of waitAny meant that we proceeded with shutdown after the first
loop finished (potentially cutting off the other two). The approach was
also a little convoluted. The new approach races a timeout against a
"wait-all", which is clearer and (hopefully) correct.